+++ both/controllers/models/FacilitySchema.jsx
@@ -100,6 +100,7 @@ FacilitySchema = {
     name: {
     	label: "Name",
     	defaultValue: "",
+        autoFocus:true
     },
     type: {
     	label:"Property type",