+++ packages/doc-thumbs/FileView.jsx
@@ -1,6 +1,7 @@
 import React from "react";
 import ReactDom from "react-dom";
 import {ReactMeteorData} from 'meteor/react-meteor-data';
+import CircularProgress from 'material-ui/CircularProgress';
 
 File = React.createClass({
 
@@ -91,21 +92,14 @@ File = React.createClass({
 
 					{this.data.file&&!this.data.file.isUploaded()?
 						<div style={{width:"100%",overflow:"hidden",textAlign:"center"}}>
-							<input 
-								ref="progress" 
-								type="text" 
-								defaultValue={0}
-								data-max={100} 
-								className="dial m-r-sm" 
-								data-fgcolor="#3ca773"
-								data-width="80" 
-								data-height="80" 
-							/>
-						</div>:<div>
-					    <img style={{width:"100%","borderRadius":"1px"}} alt="image" src={this.data.file?this.data.url:"img/default-placeholder.png"} />
-						<div style={{width:0,height:0,overflow:"hidden"}}>
-							<input ref="input" type="file" onChange={this.handleChange}/>
-						</div>
+					    	<CircularProgress />
+					    </div>
+					    :
+					    <div>
+						    <img style={{width:"100%","borderRadius":"1px"}} alt="image" src={this.data.file?this.data.url:"img/default-placeholder.png"} />
+							<div style={{width:0,height:0,overflow:"hidden"}}>
+								<input ref="input" type="file" onChange={this.handleChange}/>
+							</div>
 						</div>
 					}
 				</div>